moved helper into utils folder
