+++ src/js/abstract/utils/component-morph.js
