+++ src/js/abstract/utils/is-same-node-once.js
