+++ src/js/abstract/utils/lifecycle-logger.js
