+++ src/js/abstract/utils/property-exists-exception.js
